+++ native/c/zutm.c
@@ -8,13 +8,13 @@
  * Copyright Contributors to the Zowe Project.
  *
  */
-
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 #include "zmetal.h"
 #include "asasymbp.h"
 #include "zstorage.h"
+#include "ztype.h"
 #include "zutm.h"
 #include "ztime.h"
 #include "zutm31.h"
@@ -226,12 +226,21 @@ int ZUTSYMBP(SYMBOL_DATA *data)
 typedef int (*ISRSUPC)(void *) ATTRIBUTE(amode31);
 #pragma prolog(ZUTSRCH, " ZWEPROLG NEWDSA=(YES,4) ")
 #pragma epilog(ZUTSRCH, " ZWEEPILG ")
-
 typedef struct
 {
   short len;
   char parms[100];
 } PARMS;
+
+typedef struct
+{
+  unsigned int params[5];
+} PARAM_CHAIN31;
+
+typedef struct
+{
+  unsigned long params[5];
+} PARAM_CHAIN64;
 
 int ZUTSRCH(const char *parms)
 {
@@ -249,11 +258,13 @@ int ZUTSRCH(const char *parms)
 
 #pragma prolog(ZUTRUN, " ZWEPROLG NEWDSA=(YES,16),LOC24=YES ")
 #pragma epilog(ZUTRUN, " ZWEEPILG ")
-typedef int (*PGM31)(void *) ATTRIBUTE(amode31);
-typedef int (*PGM64)(void *) ATTRIBUTE(amode64);
 
-int ZUTRUN(ZDIAG *diag, const char *program, const char *parms)
+typedef int (*PGM31)(PARAM_CHAIN31) ATTRIBUTE(amode31);
+typedef int (*PGM64)(PARAM_CHAIN64) ATTRIBUTE(amode64);
+
+int ZUTRUN(ZDIAG *diag, const char *program, const char *parms, PROGRAM_OPTION_LIST *opt_list)
 {
+
   int rc = 0;
 
   PARMS pstruct = {0};
@@ -262,29 +273,65 @@ int ZUTRUN(ZDIAG *diag, const char *program, const char *parms)
     pstruct.len = sprintf(pstruct.parms, "%.*s", (int)(sizeof(pstruct.parms) - 1), parms);
   }
 
-  PARMS *pptr = &pstruct;
+  unsigned int opt_count = opt_list ? opt_list->count : 0;
+  if (opt_count > 4)
+  {
+    ZDIAG_SET_MSG(diag, "ZUTRUN called with more than 4 program options. Only up to 4 are supported.");
+    diag->detail_rc = ZUT_CALLER_ERROR;
+    return RTNCD_FAILURE;
+  }
 
   char name_truncated[8 + 1] = {0};
   memset(name_truncated, ' ', sizeof(name_truncated) - 1);                                                                      // pad with spaces
   memcpy(name_truncated, program, strlen(program) > sizeof(name_truncated) - 1 ? sizeof(name_truncated) - 1 : strlen(program)); // truncate
 
+  // parms[0] holds the PARMS struct; options occupy slots 1..opt_count.
+  // Re-allocation here on the stack is only technically necessary for 31 bit programs.
+  //   64-bit programs could use memory as-is. However, most program calls will be 31 bit...
+  char option_data[4][512] = {{0}};
+  PARAM_CHAIN31 pchain = {0};
+  pchain.params[0] = (unsigned int)(uintptr_t)&pstruct;
+  for (unsigned int i = 0; i < opt_count; i++)
+  {
+    PROGRAM_OPTION *p = opt_list->options[i];
+    if (p->size > sizeof(option_data[0]))
+    {
+      ZDIAG_SET_MSG(diag, "ZUTRUN option %u size %u exceeds maximum %u", i, p->size, (unsigned int)sizeof(option_data[0]));
+      diag->detail_rc = ZUT_CALLER_ERROR;
+      return RTNCD_FAILURE;
+    }
+    memcpy(option_data[i], p->data, p->size);
+    pchain.params[i + 1] = (unsigned int)(uintptr_t)option_data[i] + p->chain_ptr_offset;
+  }
+
   void *p = load_module(name_truncated);
+  // https://www.ibm.com/docs/en/zos/3.1.0?topic=c-metal-mvs-linkage-conventions
   if (p)
   {
-
     long long unsigned int ifunction = (long long unsigned int)p;
 
     if (ifunction & 0x00000000000000001ULL)
     {
       ifunction &= 0xFFFFFFFFFFFFFFFEULL; // clear low bit
+
+      // must widen to 8 bytes for 64 bit chain-of-pointers
+      PARAM_CHAIN64 pchain64 = {0};
+      for (int j = 0; j < 5; j++)
+      {
+        pchain64.params[j] = (unsigned long)pchain.params[j];
+      }
       PGM64 p64 = (PGM64)ifunction;
-      rc = p64(pptr);
+      rc = p64(pchain64);
     }
     else
     {
       ifunction &= 0x000000007FFFFFFFULL; // clear high bit
+      if (opt_count > 0)
+      { // only set high order bit when variable length opts
+        pchain.params[opt_count] |= 0x80000000U;
+      }
       PGM31 p31 = (PGM31)ifunction;
-      rc = p31(pptr);
+      rc = p31(pchain);
     }
   }
   else